bookmark listing feature added

+++ SWE574/core/views.py
@@ -8,7 +8,7 @@ from django.shortcuts import render, redirect
 from django.contrib.auth.models import auth
 from .models import Profile, Tag, Space, Post, User, Bookmark
 from .src.models import post_model_handler, tag_model_handler, space_model_handler
-from .src.models.post_model_handler import __delete_post__, __book_post__
+from .src.models.post_model_handler import __delete_post__
 from .src.pages.profile_page_handler import profile_page_handler_main
 from .src.pages.settings_page_handler import settings_page_handler_main
 from .src.pages.signin_page_handler import signin_page_handler_main
@@ -48,11 +48,6 @@ def delete_account(request: object):
 @login_required(login_url='core:signin')
 def delete_post(request: object):
     return __delete_post__(request=request)
-
-
-@login_required(login_url='core:signin')
-def book_post(request: object):
-    return __book_post__(request=request)
 
 
 @login_required(login_url="core:signin")
@@ -335,51 +330,40 @@ def dislike_post(request, post_id):
 
 
 @login_required(login_url="core:signin")
-def bookmark_post(request, post_id):
-    print("bookmark_post called")
+def bookmark_post(request):
+    data = json.loads(request.body)
+    post_id = data.get("postId")
+    label_name = data.get("labelName")
     post = Post.objects.get(post_id=post_id)
     user = User.objects.get(username=request.user.username)
-    label = request.GET.get('label')
-    print(f"{label=}")
-    # post.bookmarks.add(user)
-    bookmark = Bookmark.objects.create(
-        label=label if label else 'default',
-        post=post,
-        user=user
-    )
-    bookmark.save()
-    bookmarked = True
+    if post.bookmarks.filter(id=request.user.id).exists():
+        bookmark = Bookmark.objects.get(post=post, user=user)
+        bookmark.delete()
+        bookmarked = False
+    else:
+        bookmark = Bookmark.objects.create(
+            post=post,
+            user=user,
+            label_name=label_name
+        )
+        bookmark.save()
+        bookmarked = True
 
     return JsonResponse({'bookmarked': bookmarked})
-
-
-@login_required(login_url="core:signin")
-def debookmark_post(request, post_id):
-    print("debookmark_post called")
-    post = Post.objects.get(post_id=post_id)
-    bookmark = Bookmark.objects.get(post=post)
-    bookmark.delete()
-    bookmarked = False
-    return JsonResponse({'bookmarked': bookmarked})
-
-
-# this methods fetch the given url's og tags and return json response as img, title, description
 
 
 @login_required(login_url="core:signin")
 def fetch_og_tags(request):
+    # This methods fetch the given url's og tags and return json response as img, title, description
     """ get url from request body, post request """
     data = json.loads(request.body.decode("utf-8"))
     url = data["url"]
-    # print(url)
     response = requests.get(url)
     soup = BeautifulSoup(response.text, 'html.parser')
     og_image = soup.find('meta', property='og:image')
     og_title = soup.find('meta', property='og:title')
     og_description = soup.find('meta', property='og:description')
-    # print(og_image['content'])
-    # print(og_title['content'])
-    # print(og_description['content'])
+
     return JsonResponse({'img': og_image['content'], 'title': og_title['content'],
                          'description': og_description['content']})
